penv/requirements.txt
tensorflow-2.10.0  #TensorFlow is an open source machine learning framework for everyone
tflearn 0.5.0  #Deep Learning Library featuring a higher-level API for TensorFlow
nltk 3.7 #Natural Language Toolkit